Refactor kmeans.tsx into multiple modules
