+++ src/modules/stateSpaces/kmeans/index.ts
@@ -1,0 +1,7 @@
+import InteractiveClusteringExample from './interactiveClusteringExample';
+import KMeansStepExample from './kmeansStepExample';
+
+import KMeansDemo from './kmeans';
+
+export {InteractiveClusteringExample, KMeansStepExample};
+export default KMeansDemo;
